Move Document.id_from_manifest_uri to utils module ref #1238
